fas_api/requirements.txt
certifi==2018.8.13
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-cors-middleware==1.3.1
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-redis==4.9.0
django-registration-redux==2.4
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
oauthlib==2.1.0
pipdeptree==0.13.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
uritemplate==3.0.0
urllib3==1.23
